fix admin productos

+++ js/admin-productos.js
@@ -1,22 +1,29 @@
-function mostrarProductos(){
-    var cuerpo = document.getElementById("cuerpoTabla");
-    if(cuerpo == null) return;
-    cuerpo.innerHTML = "";
-    for(var i=0; i < productos.length; i++){
-        var p = productos[i];
-        cuerpo.innerHTML += "<tr><td>"+p.id+"</td><td>"+p.nombre+"</td><td>$"+p.precio+"</td><td>"+p.stock+"</td><td><button onclick='eliminar("+i+")'>Eliminar</button> <button onclick='editar("+i+")'>Editar</button></td></tr>";
-    }
+var guardado = localStorage.getItem("productos");
+if(guardado){
+  productos = JSON.parse(guardado);
 }
 
-function eliminar(indice){
-    productos.splice(indice, 1);
+var tabla = document.getElementById("tabla-productos");
+tabla.innerHTML = "";
+
+for(var i=0; i<productos.length; i++){
+  tabla.innerHTML +=
+  "<tr><td>"+productos[i].id+"</td>"+
+  "<td>"+productos[i].nombre+"</td>"+
+  "<td>"+productos[i].precio+"</td>"+
+  "<td>"+productos[i].stock+"</td>"+
+  "<td><button onclick='editar("+i+")'>Editar</button> <button onclick='eliminar("+i+")'>Eliminar</button></td></tr>";
+}
+
+function editar(i){
+  localStorage.setItem("indiceEditar", i);
+  window.location.href = "producto-editar.html";
+}
+
+function eliminar(i){
+  if(confirm("¿Eliminar?")){
+    productos.splice(i,1);
     localStorage.setItem("productos", JSON.stringify(productos));
-    mostrarProductos();
+    location.reload();
+  }
 }
-
-function editar(indice){
-    localStorage.setItem("indiceEditar", indice);
-    window.location.href = "producto-editar.html";
-}
-
-mostrarProductos();
